타입에러 수정 및 기능 추가

+++ app/galleryPost/page.tsx
@@ -1,13 +1,19 @@
 // galleryPost 공식레시피 상세보기
 'use client'; // 이 줄을 추가하여 이 파일이 클라이언트 컴포넌트임을 명시합니다.
 
-import React, { useEffect, useState } from 'react'; // React와 필요한 훅들을 import
+import React, {
+    CSSProperties,
+    HTMLProps,
+    ReactNode,
+    useEffect,
+    useState,
+} from 'react'; // React와 필요한 훅들을 import
 import { useRouter } from 'next/navigation'; // Next.js의 라우팅 훅을 import
 import Image from 'next/image'; // 이미지 최적화를 위한 Next.js Image 컴포넌트
 import xml2js from 'xml2js'; // XML 데이터를 파싱하기 위한 라이브러리
 import { FaArrowLeft, FaPause, FaPlay, FaStop, FaHeart } from 'react-icons/fa'; // 아이콘을 위한 라이브러리
 import { db, auth } from '../../lib/firebaseConfig'; // Firebase 설정 파일 import
-import { onAuthStateChanged } from 'firebase/auth'; // Firebase 인증 상태 변화를 감지하는 함수
+import { onAuthStateChanged, User } from 'firebase/auth'; // Firebase 인증 상태 변화를 감지하는 함수
 import {
     query,
     where,
@@ -16,9 +22,16 @@ import {
     getDocs,
     deleteDoc,
 } from 'firebase/firestore'; // Firestore 관련 함수들 import
+import { StaticImport } from 'next/dist/shared/lib/get-img-props';
 
-// 임시 Box, Text 컴포넌트
-const Box = ({ children, style, ...props }) => (
+////////////// 임시 Box, Text 컴포넌트 //////////////
+interface BoxProps {
+    children: ReactNode;
+    style?: CSSProperties;
+    [key: string]: any;
+}
+
+const Box: React.FC<BoxProps> = ({ children, style, ...props }) => (
     <div
         style={style}
         {...props}
@@ -27,7 +40,12 @@ const Box = ({ children, style, ...props }) => (
     </div>
 );
 
-const Text = ({ children, style, ...props }) => (
+interface TextProps extends HTMLProps<HTMLParagraphElement> {
+    children: ReactNode;
+    style?: CSSProperties;
+}
+
+const Text: React.FC<TextProps> = ({ children, style, ...props }) => (
     <p
         style={style}
         {...props}
@@ -35,19 +53,46 @@ const Text = ({ children, style, ...props }) => (
         {children}
     </p>
 );
+/////////////////////////////////////////////////////////////////////////////////
+
+// 타입 정의
+interface Recipe {
+    calories: ReactNode;
+    protein: ReactNode;
+    fat: ReactNode;
+    sodium: ReactNode;
+    ingredients: ReactNode;
+    image: string | StaticImport;
+    manual: any;
+    name: string;
+    id: string;
+    // 기타 속성들
+}
+interface Comment {
+    userId: string;
+    userEmail: string | null;
+    // userName: string;
+    text: string;
+    recipeId: string;
+}
 
 const GalleryPost = () => {
     const router = useRouter(); // 페이지 이동을 위한 라우터 객체 생성
-    const [recipe, setRecipe] = useState(null); // 레시피 데이터를 저장할 상태 변수
+    const [recipe, setRecipe] = useState<Recipe | null>(null); // 레시피 데이터를 저장할 상태 변수
     const [loading, setLoading] = useState(true); // 데이터 로딩 상태를 관리하는 변수
     const [speechRate, setSpeechRate] = useState(1); // TTS 재생 속도를 관리하는 상태 변수
-    const [utterance, setUtterance] = useState(null); // TTS Utterance 객체를 저장할 상태 변수
-    const [isSpeaking, setIsSpeaking] = useState(false); // 현재 TTS가 재생 중인지 여부를 저장하는 상태 변수
-    const [liked, setLiked] = useState(false); // 현재 사용자가 좋아요를 눌렀는지 여부를 저장하는 상태 변수
+    const [utterance, setUtterance] = useState<SpeechSynthesisUtterance | null>(
+        null
+    ); // TTS Utterance 객체를 저장할 상태 변수
+    const [isSpeaking, setIsSpeaking] = useState<boolean>(false); // 현재 TTS가 재생 중인지 여부를 저장하는 상태 변수
+    const [isPaused, setIsPaused] = useState<boolean>(false); // 현재 TTS가 일시 정지 상태인지 여부를 저장하는 상태 변수
+    const [liked, setLiked] = useState<boolean>(false); // 현재 사용자가 좋아요를 눌렀는지 여부를 저장하는 상태 변수
     const [newComment, setNewComment] = useState(''); // 새 댓글 내용을 저장할 상태 변수
-    const [comments, setComments] = useState([]); // 댓글 목록을 저장할 상태 변수
-    const [user, setUser] = useState(null); // 현재 로그인한 사용자 정보를 저장할 상태 변수
+    const [comments, setComments] = useState<Comment[]>([]); // 댓글 목록을 저장할 상태 변수
+    const [user, setUser] = useState<User | null>(null); // 현재 로그인한 사용자 정보를 저장할 상태 변수
     const [searchTerm, setSearchTerm] = useState(''); // 검색어를 저장할 상태 변수
+
+    // 상태 타입을 SpeechSynthesisUtterance | null로 설정
 
     // 레시피 데이터를 가져오는 useEffect
     useEffect(() => {
@@ -63,7 +108,7 @@ const GalleryPost = () => {
 
                 const recipes = result.COOKRCP01.row; // 레시피 목록을 가져옴
                 const foundRecipe = recipes.find(
-                    (rec) => rec.RCP_SEQ[0] === recipeId
+                    (rec: any) => rec.RCP_SEQ[0] === recipeId
                 ); // URL에서 추출한 레시피 ID와 일치하는 레시피를 찾음
 
                 if (foundRecipe) {
@@ -93,7 +138,7 @@ const GalleryPost = () => {
                         fat: foundRecipe.INFO_FAT[0],
                         sodium: foundRecipe.INFO_NA[0],
                     };
-                    setRecipe(recipeData); // 상태 변수에 레시피 데이터 저장
+                    setRecipe(recipeData as any); // 상태 변수에 레시피 데이터 저장
                 }
             } catch (error) {
                 console.error('Error parsing XML:', error); // 에러 발생 시 콘솔에 출력
@@ -114,7 +159,6 @@ const GalleryPost = () => {
         return () => unsubscribe(); // 컴포넌트 언마운트 시 구독 해제
     }, []);
 
-    // 사용자가 이미 좋아요를 눌렀는지 확인하는 useEffect
     useEffect(() => {
         const checkIfLiked = async () => {
             if (user && recipe) {
@@ -133,8 +177,10 @@ const GalleryPost = () => {
         checkIfLiked();
     }, [user, recipe]); // user와 recipe 상태가 변경될 때마다 실행
 
+    //////////////////////////////////////// TTS 관련 ///////////////////////////////////////////
+
     // 텍스트를 음성으로 읽어주는 함수
-    const speakText = (text) => {
+    const speakText = (text: string) => {
         if ('speechSynthesis' in window) {
             const newUtterance = new SpeechSynthesisUtterance(text); // 새 Utterance 객체 생성
             newUtterance.rate = speechRate; // TTS 속도 설정
@@ -151,28 +197,29 @@ const GalleryPost = () => {
         if (recipe) {
             // 레시피 이름과 메뉴얼 텍스트를 음성으로 읽어줌
             speakText(recipe.name);
-            recipe.manual.forEach((item) => {
+            recipe.manual.forEach((item: { text: string }) => {
                 speakText(item.text);
             });
         }
     };
 
     // TTS 속도 변경 시 호출되는 함수
-    const handleRateChange = (event) => {
+    const handleRateChange = (event: any) => {
         setSpeechRate(parseFloat(event.target.value)); // 선택된 속도로 업데이트
     };
 
-    // TTS 일시정지 버튼 클릭 시 호출되는 함수
-    const handlePauseClick = () => {
-        if ('speechSynthesis' in window && isSpeaking) {
-            speechSynthesis.pause(); // TTS 일시정지
-        }
-    };
-
-    // TTS 재개 버튼 클릭 시 호출되는 함수
-    const handleResumeClick = () => {
-        if ('speechSynthesis' in window && isSpeaking) {
-            speechSynthesis.resume(); // TTS 재개
+    // TTS 일시정지 및 재개 함수
+    const handlePauseOrResumeClick = () => {
+        if ('speechSynthesis' in window) {
+            if (isSpeaking) {
+                if (isPaused) {
+                    speechSynthesis.resume(); // TTS 재개
+                    setIsPaused(false);
+                } else {
+                    speechSynthesis.pause(); // TTS 일시 정지
+                    setIsPaused(true);
+                }
+            }
         }
     };
 
@@ -184,6 +231,8 @@ const GalleryPost = () => {
         }
     };
 
+    ////////////////////////////////////////////////////////////////////////////////////////////
+
     // 뒤로가기 버튼 클릭 시 호출되는 함수
     const handleBackClick = () => {
         router.back(); // 이전 페이지로 이동
@@ -191,68 +240,83 @@ const GalleryPost = () => {
 
     // 좋아요 버튼 클릭 시 호출되는 함수
     const handleLikeToggle = async () => {
-        if (user) {
+        const router = useRouter();
+
+        if (user && recipe) {
             const likesRef = collection(db, 'likes');
             const q = query(
                 likesRef,
                 where('recipeId', '==', recipe.id),
                 where('userId', '==', user.uid)
             );
-            const querySnapshot = await getDocs(q);
 
-            if (querySnapshot.empty) {
-                // 좋아요가 없는 경우, 좋아요 추가
-                try {
+            try {
+                const querySnapshot = await getDocs(q);
+
+                if (querySnapshot.empty) {
+                    // 좋아요가 없는 경우, 좋아요 추가
                     await addDoc(likesRef, {
                         recipeId: recipe.id,
                         userId: user.uid,
                     });
                     setLiked(true);
-                } catch (error) {
-                    console.error('Error adding like:', error);
-                }
-            } else {
-                // 좋아요가 있는 경우, 좋아요 제거
-                try {
-                    querySnapshot.forEach(async (doc) => {
-                        await deleteDoc(doc.ref);
-                    });
+                } else {
+                    // 좋아요가 있는 경우, 좋아요 제거
+                    const deletePromises = querySnapshot.docs.map((doc) =>
+                        deleteDoc(doc.ref)
+                    );
+                    await Promise.all(deletePromises);
                     setLiked(false);
-                } catch (error) {
-                    console.error('Error removing like:', error);
                 }
+            } catch (error) {
+                console.error('Error handling like:', error);
             }
         } else {
             // 로그인하지 않은 경우 알림과 리다이렉트
-            alert('좋아요를 클릭하려면 로그인해야 합니다.');
-            router.push('/login'); // 로그인 페이지로 리다이렉트
+            if (!user) {
+                alert('좋아요를 클릭하려면 로그인해야 합니다.');
+                router.push('/login'); // 로그인 페이지로 리다이렉트
+            }
+            // recipe가 null인 경우에 대한 추가 처리 로직이 필요한 경우 여기에 작성
+            if (recipe === null) {
+                return <div>레시피를 찾을 수 없습니다.</div>;
+            }
         }
     };
 
     // 새 댓글을 추가하는 함수
     const handleAddComment = async () => {
         if (user) {
-            if (newComment.trim()) {
+            if (recipe && newComment.trim()) {
                 try {
                     await addDoc(collection(db, 'comments'), {
                         recipeId: recipe.id,
                         userId: user.uid,
-                        userEmail: user.email, // 이메일 주소 추가
+                        userEmail: user.email,
                         text: newComment,
                         timestamp: new Date(),
                     });
-                    setComments([
-                        ...comments,
+
+                    // 댓글 추가
+                    setComments((prevComments) => [
+                        ...prevComments,
                         {
                             userId: user.uid,
                             userEmail: user.email,
+                            // userName: user.name,
                             text: newComment,
+                            recipeId:recipe.id
                         },
                     ]);
+
                     setNewComment(''); // 댓글 입력 필드 초기화
                 } catch (error) {
                     console.error('Error adding comment:', error);
                 }
+            } else {
+                console.warn(
+                    '댓글 내용이 비어 있거나 레시피가 선택되지 않았습니다.'
+                );
             }
         } else {
             // 로그인하지 않은 경우 알림과 리다이렉트
@@ -266,10 +330,22 @@ const GalleryPost = () => {
         const fetchComments = async () => {
             try {
                 const querySnapshot = await getDocs(collection(db, 'comments'));
-                const commentsList = querySnapshot.docs.map((doc) =>
-                    doc.data()
-                );
-
+                
+                const commentsList: Comment[] = querySnapshot.docs.map((doc) => {
+                    const data = doc.data() as Comment;
+    
+                    // 데이터 검증 (선택 사항)
+                    if (typeof data.userId === 'string' &&
+                        (typeof data.userEmail === 'string' || data.userEmail === null) &&
+                        typeof data.text === 'string' &&
+                        typeof data.recipeId === 'string') {
+                        return data;
+                    } else {
+                        console.warn('Invalid comment data:', data);
+                        return null; // 필터링하여 무시할 수 있음
+                    }
+                }).filter((comment): comment is Comment => comment !== null); // null 제외
+    
                 // 댓글 목록에서 recipeId가 일치하는 댓글만 필터링
                 setComments(
                     commentsList.filter(
@@ -280,19 +356,16 @@ const GalleryPost = () => {
                 console.error('Error fetching comments:', error);
             }
         };
-
+    
         if (recipe) {
             fetchComments(); // 레시피가 로드된 후 댓글 데이터 가져오기
         }
     }, [recipe]);
 
-    const handleSearch = () => {
-        const filteredRecipes = recipes.filter((recipe) =>
-            recipe.name.toLowerCase().includes(searchTerm.toLowerCase())
-        );
-        setFilteredRecipes(filteredRecipes); // 필터링된 레시피를 상태로 설정
-    };
 
+
+
+    
     return (
         <main style={{ marginTop: '80px' }}>
             <Box style={{ padding: '16px' }}>
@@ -410,7 +483,7 @@ const GalleryPost = () => {
                             >
                                 조리법
                             </h2>
-                            {recipe.manual.map((item, index) => (
+                            {recipe.manual.map((item: any, index: any) => (
                                 <div
                                     key={index}
                                     style={{
@@ -495,14 +568,16 @@ const GalleryPost = () => {
                                     }}
                                 >
                                     <option value={0.5}>0.5배속</option>
-                                    <option value={1}>1배속</option>
-                                    <option value={2}>2배속</option>
+                                    <option value={1.5}>1배속</option>
+                                    <option value={2.5}>2배속</option>
                                 </select>
                                 <button
-                                    onClick={handlePauseClick} // TTS 일시정지 클릭 시 handlePauseClick 호출
+                                    onClick={handlePauseOrResumeClick}
                                     style={{
                                         color: '#ffffff',
-                                        backgroundColor: '#FF8C00', // 기본 주황색
+                                        backgroundColor: isPaused
+                                            ? '#FF7F00'
+                                            : '#FF8C00', // 상태에 따라 색상 변경
                                         width: 80,
                                         height: 40,
                                         borderRadius: 20,
@@ -518,47 +593,20 @@ const GalleryPost = () => {
                                     }}
                                     onMouseDown={(e) => {
                                         e.currentTarget.style.backgroundColor =
-                                            '#FF7F00'; // 클릭 시 색상
+                                            isPaused ? '#FF8C00' : '#FF7F00'; // 클릭 시 색상
                                         e.currentTarget.style.color = '#ffffff'; // 클릭 시 글씨 색상
                                     }}
                                     onMouseUp={(e) => {
                                         e.currentTarget.style.backgroundColor =
-                                            '#FF8C00'; // 기본 색상으로 복구
+                                            isPaused ? '#FF7F00' : '#FF8C00'; // 기본 색상으로 복구
                                         e.currentTarget.style.color = '#ffffff'; // 기본 글씨 색상
                                     }}
                                 >
-                                    <FaPause size={15} />
-                                </button>
-                                <button
-                                    onClick={handleResumeClick} // TTS 재개 클릭 시 handleResumeClick 호출
-                                    style={{
-                                        color: '#ffffff',
-                                        backgroundColor: '#FF8C00', // 기본 주황색
-                                        width: 80,
-                                        height: 40,
-                                        borderRadius: 20,
-                                        border: 'none',
-                                        display: 'flex',
-                                        alignItems: 'center',
-                                        justifyContent: 'center',
-                                        cursor: 'pointer',
-                                        fontSize: '16px',
-                                        boxShadow: '0 4px 8px rgba(0,0,0,0.1)',
-                                        transition:
-                                            'background-color 0.3s, box-shadow 0.3s, color 0.3s',
-                                    }}
-                                    onMouseDown={(e) => {
-                                        e.currentTarget.style.backgroundColor =
-                                            '#FF7F00'; // 클릭 시 색상
-                                        e.currentTarget.style.color = '#ffffff'; // 클릭 시 글씨 색상
-                                    }}
-                                    onMouseUp={(e) => {
-                                        e.currentTarget.style.backgroundColor =
-                                            '#FF8C00'; // 기본 색상으로 복구
-                                        e.currentTarget.style.color = '#ffffff'; // 기본 글씨 색상
-                                    }}
-                                >
-                                    <FaPlay size={15} />
+                                    {isPaused ? (
+                                        <FaPlay size={15} />
+                                    ) : (
+                                        <FaPause size={15} />
+                                    )}
                                 </button>
                                 <button
                                     onClick={handleStopClick} // TTS 정지 클릭 시 handleStopClick 호출
